Update maintainers guide
This makes some changes to the maintainer's guide and roles within the
project.  It removed the concept of a BDFL and carries over the chief
maintainer role into the project.  BDFL sucks and we can do better and
these changed help to make many more things specific around adding new
maintainers and removing them as well.

Signed-off-by: Michael Crosby <crosbymichael@gmail.com>