rootless: fix running with /proc/self/setgroups set to deny
This is a regression from aa09b40da01d0d36cef7edffa86c291c282aadab
when the user namespace was configured without a privileged helper.
To allow a single mapping in an user namespace, it is necessary to set
/proc/self/setgroups to "deny".

For a simple reproducer, the user namespace can be created with
"unshare -r".

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>