Add support for uid shifting to sysvisor-runc.
This change adds support for uid shifting to sysvisor-runc. Note
that uid shifting needs the presence of the shiftfs module in
the kernel.

With uid shifting, sysvisor-runc can now run system containers whose
rootfs or volume bind mounts are owned by true root; this is not
possible without uid shifting as the container's root process, which
maps to a regular user on the host, would normally not have access to
such directories.

This feature makes it possible to run system containers without
using docker userns-remap. It also makes it possible to mount
shared volumes owned by true root in system containers.

With this change, sysvisor-runc automatically detects the cases
where it must use uid shifting and when it must not. This makes
it possible to use sysvisor-runc with or without docker userns
remap, without any extra configuration.

To make this feature possible, this change reworks the procedure
used by the container's init process to initialize the container's
mounts.

Prior to this change, the init process did all the mounts and used
absolute paths to perform them. With this change, the init process
only does mounts inside the container rootfs and uses paths relatives
to the rootfs. This way the init process need not access paths
outside the rootfs since it may not have access to them.

When uid shifting is required, the first mount done by the init
process is to mount shiftfs on the container's rootfs. It then
performs other mounts inside the container's rootfs.

When performing bind mounts whose source is outside of the container's
rootfs, the init process requests the parent sysvisor-runc process to
perform the mount on its behalf. The sysvisor-runc process does
this by spawning a child helper process that enters the container's
mount namespace but not it's user namespace. This way, the helper
process can perform the mount in the correct mount namespace, yet
has credentials to perform the mount (since it can access the
source of the bind mount).

This change also adds code to perform a dummy allocation of uid/gid
mappings when these are not present in the container spec.

Finally, several fixes where done to ensure sysvisor-runc unit tests
pass with and without uid shifting.