Ignore AppArmor profile in the container's config.
This change causes sysbox-runc to ignore the AppArmor profile
in the container's config.

The change was required because in systems where AppArmor is enabled,
Docker configures containers with a default profile that is too
restrictive for system containers and prevents them from working
properly (e.g., mounts are not possible, access to /proc/sys/*
is restricted, etc.)

In the future it is envisiones that sysbox will have a default
profile for sys containers and it will be used instead.