Re-work mount sequence to avoid shiftfs-on-shiftfs mounts and allow shiftfs-on-tmpfs.
This change re-works the sysvisor-runc sequence for mounting shiftfs
on bind sources to mounting shiftfs above a directory on which shiftfs
is already mounted. This is necessary as mounting shiftfs-on-shiftfs
does not work because the uid shifting is done twice, resulting in
incorrect operation.

In addition, this change removes code that was disallowing
shiftfs-on-tmpfs mounts, as Docker uses a bind mount to a tmpfs
directory for the container's /dev/shm.