Modularize code that interacts with sysvisor-mgr and sysvisor-fs.
This reduces pollution of libcontainer with such code, and in addition
allows sysvisor-runc to keep state regarding the interactions with
sysvisor-mgr and sysvisor-fs; such state is needed during container
creation and destruction.

This change also fixes a bug: sysvisor-runc now releases sysvisor-fs
and sysvisor-mgr resources when container fails to launch. Prior to
this change it wasn't doing this.

Finally, this change adds code in libsysvisor to obtain supplementary
mounts from sysvisor-mgr. This sets the foundation for upcoming
changes related to sysvisor epic #46 (https://github.com/nestybox/sysvisor/issues/46).