Modularize code that interacts with sysbox-mgr and sysbox-fs.
This reduces pollution of libcontainer with such code, and in addition
allows sysbox-runc to keep state regarding the interactions with
sysbox-mgr and sysbox-fs; such state is needed during container
creation and destruction.

This change also fixes a bug: sysbox-runc now releases sysbox-fs
and sysbox-mgr resources when container fails to launch. Prior to
this change it wasn't doing this.

Finally, this change adds code in libsysbox to obtain supplementary
mounts from sysbox-mgr.