Fix ownership of container init process stdio in detached mode.
The container's stdio ownership must match the uid(gid) of the
contaner's init process.

Prior to this change, this was not the case when launching sys
containers with docker (without userns-remap). That's because
containerd-shim was setting the stdio ownership to root (since Docker
thinks the container will have uid(gid) = root), but sysvisor then
allocates non-root uid(gid) for the container.

Refer to sysvisor issue #316 for full details.

I tested that this works by running a container in detached but
interactive mode (docker run --sysvisor-runc -i -d ...) and then
attaching to the container (docker attach) and verifying that both
input and output from the container work as expected.